requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
arrow==1.3.0
asyncio==3.4.3
attrs==23.2.0
azure-core==1.30.1
binaryornot==0.4.4
botbuilder-core==4.15.0
botbuilder-integration-aiohttp==4.15.0
botbuilder-schema==4.15.0
botframework-connector==4.15.0
botframework-streaming==4.15.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==1.7.0
cryptography==42.0.7
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.3
frozenlist==1.4.1
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
isodate==0.6.1
Jinja2==3.1.4
jinja2-time==0.2.0
jsonpickle==1.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msal==1.28.0
msrest==0.7.1
multidict==6.0.5
oauthlib==3.2.2
orjson==3.10.3
poyo==0.5.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
ujson==5.9.0
urllib3==1.26.18
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
whichcraft==0.6.1
yarl==1.9.4
